chore: Moves classes to separate files to respect the 1 class per file rule. Might change this decision if it turns out to make sense when following some convention.
