+++ src/Tutor.Core/DomainModel/AssessmentEvents/ArrangeTasks/ArrangeTaskContainerEvaluation.cs
@@ -1,0 +1,14 @@
+﻿namespace Tutor.Core.DomainModel.AssessmentEvents.ArrangeTasks
+{
+    public class ArrangeTaskContainerEvaluation
+    {
+        public ArrangeTaskContainer FullAnswer { get; }
+        public bool SubmissionWasCorrect { get; }
+
+        public ArrangeTaskContainerEvaluation(ArrangeTaskContainer fullAnswer, bool submissionWasCorrect)
+        {
+            FullAnswer = fullAnswer;
+            SubmissionWasCorrect = submissionWasCorrect;
+        }
+    }
+}
