+++ src/Tutor.Core/DomainModel/AssessmentEvents/ArrangeTasks/ArrangeTaskContainerSubmission.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Tutor.Core.DomainModel.AssessmentEvents.ArrangeTasks
+{
+    public class ArrangeTaskContainerSubmission
+    {
+        public int Id { get; private set; }
+        public int ContainerId { get; private set; }
+        public List<int> ElementIds { get; private set; }
+    }
+}
