+++ src/Tutor.Core/DomainModel/AssessmentEvents/ArrangeTasks/ArrangeTaskEvaluation.cs
@@ -11,16 +11,4 @@ namespace Tutor.Core.DomainModel.AssessmentEvents.ArrangeTasks
             ContainerEvaluations = containerEvaluations;
         }
     }
-
-    public class ArrangeTaskContainerEvaluation
-    {
-        public ArrangeTaskContainer FullAnswer { get; }
-        public bool SubmissionWasCorrect { get; }
-
-        public ArrangeTaskContainerEvaluation(ArrangeTaskContainer fullAnswer, bool submissionWasCorrect)
-        {
-            FullAnswer = fullAnswer;
-            SubmissionWasCorrect = submissionWasCorrect;
-        }
-    }
 }
