+++ src/Tutor.Core/DomainModel/AssessmentEvents/ArrangeTasks/ArrangeTaskSubmission.cs
@@ -6,11 +6,4 @@ namespace Tutor.Core.DomainModel.AssessmentEvents.ArrangeTasks
     {
         public List<ArrangeTaskContainerSubmission> Containers { get; private set; }
     }
-
-    public class ArrangeTaskContainerSubmission
-    {
-        public int Id { get; private set; }
-        public int ContainerId { get; private set; }
-        public List<int> ElementIds { get; private set; }
-    }
 }
