+++ src/Tutor.Core/DomainModel/AssessmentEvents/MultiResponseQuestions/MRQEvaluation.cs
@@ -10,16 +10,4 @@ namespace Tutor.Core.DomainModel.AssessmentEvents.MultiResponseQuestions
             ItemEvaluations = evaluations;
         }
     }
-
-    public class MRQItemEvaluation
-    {
-        public MRQItem FullItem { get; }
-        public bool SubmissionWasCorrect { get; }
-
-        public MRQItemEvaluation(MRQItem item, bool isCorrect)
-        {
-            FullItem = item;
-            SubmissionWasCorrect = isCorrect;
-        }
-    }
 }
